[Chore/#6] MypageSetting - ViewModel 분리
